pyproject.toml
[project]
name = "simulative-basic-python"
version = "0.1.0"
description = ""
authors = [
    {name = "akaserpens aka.serpens@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "gspread (>=6.2.1,<7.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
